fix getting_started links fr
a tricky part of this is that the relative link needs to
work from both the top-level page (endpoint `/`) AND
`/getting_started/getting_started.html`.
So if you do a relative path to `/getting_started/quick_start.html`
as `./quick_start.md`, it will work from
`/getting_started/getting_started.html` but NOT from `/`.

however, the top-level apparently has no problem with extra `../`s,
so we just include that and give the full path and that works
(i.e. `../getting_started/getting_started.md`)